chore: add design layer for list page headers
